+++ src/IoTSharp.Data.Taos/TaosDataReader.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c)  maikebing All rights reserved.
 //// Licensed under the MIT License, See License.txt in the project root for license information.
 
+using IoTSharp.Data.Taos.Protocols;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -24,28 +25,22 @@ namespace IoTSharp.Data.Taos
         private readonly int _recordsAffected;
         private bool _closed;
         private bool _closeConnection;
-        private IntPtr _taosResult;
         private int _fieldCount;
-        IntPtr rowdata=IntPtr.Zero;
-        List<taosField> _metas = null;
-        private double _date_max_1970;
-        private DateTime _dt1970;
-
-
-
-        internal TaosDataReader(TaosCommand taosCommand, taosField[] metas, bool closeConnection, IntPtr taos, IntPtr res, int recordsAffected, int fieldcount)
+     
+        internal ITaosProtocol taos =>_command?._connection?.taos;
+        ITaosContext _context;
+        internal TaosDataReader(TaosCommand taosCommand, ITaosContext context)
         {
             _command = taosCommand;
-            _closeConnection = closeConnection;
-            _fieldCount = fieldcount;
-            _hasRows = recordsAffected > 0;
-            _recordsAffected = recordsAffected;
+            _closeConnection = context.CloseConnection;
+            _fieldCount = context.FieldCount;
+            _hasRows = context.AffectRows > 0;
+            _recordsAffected = context.AffectRows;
             _closed = _closeConnection;
-            _taosResult = res;
-            _metas =  metas?.ToList()??new List<taosField> ();
-            _dt1970 = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
-            _date_max_1970 = DateTime.MaxValue.Subtract(_dt1970).TotalMilliseconds;
+            _context=context; 
         }
+
+ 
 
         /// <summary>
         ///     Gets the depth of nesting for the current row. Always zero.
@@ -66,9 +61,9 @@ namespace IoTSharp.Data.Taos
         public override bool HasRows
             => _hasRows;
 
-        public TaosException LastTaosException => TDengine.ErrorNo(_taosResult) == 0 ? null : new TaosException(new TaosErrorResult() { Code = TDengine.ErrorNo(_taosResult), Error = TDengine.Error(_taosResult) }, null);
+        public TaosException LastTaosException => _context.LastException();
 
-        int ErrorNo => TDengine.ErrorNo(_taosResult);
+        int ErrorNo => _context.GetErrorNo();
         /// <summary>
         ///     Gets a value indicating whether the data reader is closed.
         /// </summary>
@@ -122,10 +117,7 @@ namespace IoTSharp.Data.Taos
             {
                 throw new InvalidOperationException($"DataReaderClosed{nameof(Read)}");
             }
-
-            rowdata = TDengine.FetchRows(_taosResult);
-            return rowdata != IntPtr.Zero;
-
+            return _context.Read();
         }
 
         /// <summary>
@@ -160,17 +152,11 @@ namespace IoTSharp.Data.Taos
             _command.DataReader = null;
            
             _closed = true;
-            if (_taosResult != IntPtr.Zero)
-            {
-                TDengine.FreeResult(_taosResult);
-                _taosResult= IntPtr.Zero;
-            }
-            if (rowdata != IntPtr.Zero)
-            {
-                TDengine.FreeResult(rowdata);
-                rowdata= IntPtr.Zero;
-            }
+            _context?.Dispose();
             OnDispose?.Invoke(this, EventArgs.Empty);
+            _context = null;
+          
+        
         }
 
         /// <summary>
@@ -180,7 +166,7 @@ namespace IoTSharp.Data.Taos
         /// <returns>The name of the column.</returns>
         public override string GetName(int ordinal)
         {
-            return _metas[ordinal].Name;
+            return _context.GetName(ordinal);
         }
 
         /// <summary>
@@ -188,8 +174,7 @@ namespace IoTSharp.Data.Taos
         /// </summary>
         /// <param name="name">The name of the column.</param>
         /// <returns>The zero-based column ordinal.</returns>
-        public override int GetOrdinal(string name)
-            => _metas.IndexOf(_metas.FirstOrDefault(m => m.Name == name));
+        public override int GetOrdinal(string name) =>_context.GetOrdinal(name);
 
         public override string GetDataTypeName(int ordinal)
         {
@@ -201,14 +186,7 @@ namespace IoTSharp.Data.Taos
         /// </summary>
         /// <param name="ordinal">The zero-based column ordinal.</param>
         /// <returns>The data type of the column.</returns>
-        public override Type GetFieldType(int ordinal)
-        {
-            if (_metas == null || ordinal >= _metas.Count)
-            {
-                throw new InvalidOperationException($"DataReaderClosed{nameof(GetFieldType)}");
-            }
-            return _metas[ordinal].CrlType;
-        }
+        public override Type GetFieldType(int ordinal) =>_context.GetFieldType(ordinal);
 
         /// <summary>
         ///     Gets a value indicating whether the specified column is <see cref="DBNull" />.
@@ -223,14 +201,14 @@ namespace IoTSharp.Data.Taos
         /// </summary>
         /// <param name="ordinal">The zero-based column ordinal.</param>
         /// <returns>The value of the column.</returns>
-        public override bool GetBoolean(int ordinal) => Marshal.ReadByte(GetValuePtr(ordinal)) == 0 ? false : true;
+        public override bool GetBoolean(int ordinal) => _context.GetBoolean(ordinal);
 
         /// <summary>
         ///     Gets the value of the specified column as a <see cref="byte" />.
         /// </summary>
         /// <param name="ordinal">The zero-based column ordinal.</param>
         /// <returns>The value of the column.</returns>
-        public override byte GetByte(int ordinal) => Marshal.ReadByte(GetValuePtr(ordinal));
+        public override byte GetByte(int ordinal) => _context. GetByte(ordinal);
 
         /// <summary>
         ///     Gets the value of the specified column as a <see cref="char" />.
@@ -246,7 +224,7 @@ namespace IoTSharp.Data.Taos
         /// <returns>The value of the column.</returns>
         public override DateTime GetDateTime(int ordinal)
         {
-            return GetDateTimeFrom(GetValuePtr(ordinal));
+            return _context.GetDataTime(ordinal);
         }
 
         /// <summary>
@@ -264,28 +242,8 @@ namespace IoTSharp.Data.Taos
         /// </summary>
         /// <param name="ordinal">The zero-based column ordinal.</param>
         /// <returns>The value of the column.</returns>
-        public virtual TimeSpan GetTimeSpan(int ordinal)
-        {
-            var val = Marshal.ReadInt64(GetValuePtr(ordinal));
-            var _dateTimePrecision = (TSDB_TIME_PRECISION)TDengine.ResultPrecision(_taosResult);
-            switch (_dateTimePrecision)
-            {
-                /*
-                * ticks为100纳秒，必须乘以10才能达到微秒级的区分度
-                * 1秒s    = 1000毫秒ms
-                * 1毫秒ms = 1000微秒us
-                * 1微秒us = 1000纳秒ns
-                * 因此， 1毫秒ms = 1000000纳秒ns = 10000ticks
-                */
-                case TSDB_TIME_PRECISION.TSDB_TIME_PRECISION_NANO:
-                    return TimeSpan.FromTicks(val / 100);
-                case TSDB_TIME_PRECISION.TSDB_TIME_PRECISION_MICRO:
-                    return TimeSpan.FromTicks(val * 10);
-                case TSDB_TIME_PRECISION.TSDB_TIME_PRECISION_MILLI:
-                default:
-                    return TimeSpan.FromTicks(val * 10000);
-            }
-        }
+        public virtual TimeSpan GetTimeSpan(int ordinal) =>_context.GetTimeSpan(ordinal);
+     
 
         /// <summary>
         ///     Gets the value of the specified column as a <see cref="decimal" />.
@@ -377,13 +335,9 @@ namespace IoTSharp.Data.Taos
         /// <param name="bufferOffset">The index to which the data will be copied.</param>
         /// <param name="length">The maximum number of bytes to read.</param>
         /// <returns>The actual number of bytes read.</returns>
-        public override long GetBytes(int ordinal, long dataOffset, byte[] buffer, int bufferOffset, int length)
-        {
-            byte[] buffer1 = new byte[length + bufferOffset];
-            Marshal.Copy(GetValuePtr(ordinal), buffer1, (int)dataOffset, length + bufferOffset);
-            Array.Copy(buffer1, bufferOffset, buffer, 0, length);
-            return length;
-        }
+        public override long GetBytes(int ordinal, long dataOffset, byte[] buffer, int bufferOffset, int length) 
+                    => _context.GetBytes(ordinal, dataOffset, buffer, bufferOffset, length);
+  
 
         /// <summary>
         ///     Reads a stream of characters from the specified column. Not supported.
@@ -404,20 +358,8 @@ namespace IoTSharp.Data.Taos
         /// </summary>
         /// <param name="ordinal">The zero-based column ordinal.</param>
         /// <returns>The returned object.</returns>
-        public override Stream GetStream(int ordinal)
-        {
-            MemoryStream result = null;
-             var meta = _metas[ordinal];
-            int offset = IntPtr.Size * ordinal;
-            IntPtr data = Marshal.ReadIntPtr(rowdata, offset);
-            if (data != IntPtr.Zero)
-            {
-                byte[] bf = new byte[meta.Size];
-                Marshal.Copy(data, bf, 0, meta.Size);
-                result = new MemoryStream(bf);
-            }
-            return result;
-        }
+        public override Stream GetStream(int ordinal) => _context.GetStream(ordinal);
+      
 
         /// <summary>
         ///     Gets the value of the specified column.
@@ -427,233 +369,16 @@ namespace IoTSharp.Data.Taos
         /// <returns>The value of the column.</returns>
         public override T GetFieldValue<T>(int ordinal) => (T)Convert.ChangeType(GetValue(ordinal), typeof(T));
 
-
-        public IntPtr GetValuePtr(int ordinal)
-        {
-            int offset = IntPtr.Size * ordinal;
-            return Marshal.ReadIntPtr(rowdata, offset);
-        }
-
-        public static System.Text.Encoding GetType(FileStream fs)
-        {
-            byte[] Unicode = new byte[] { 0xFF, 0xFE, 0x41 };
-            byte[] UnicodeBIG = new byte[] { 0xFE, 0xFF, 0x00 };
-            byte[] UTF8 = new byte[] { 0xEF, 0xBB, 0xBF }; //带BOM
-            Encoding reVal = Encoding.Default;
-
-            BinaryReader r = new BinaryReader(fs, System.Text.Encoding.Default);
-            int i;
-            int.TryParse(fs.Length.ToString(), out i);
-            byte[] ss = r.ReadBytes(i);
-            if (IsUTF8Bytes(ss) || (ss[0] == 0xEF && ss[1] == 0xBB && ss[2] == 0xBF))
-            {
-                reVal = Encoding.UTF8;
-            }
-            else if (ss[0] == 0xFE && ss[1] == 0xFF && ss[2] == 0x00)
-            {
-                reVal = Encoding.BigEndianUnicode;
-            }
-            else if (ss[0] == 0xFF && ss[1] == 0xFE && ss[2] == 0x41)
-            {
-                reVal = Encoding.Unicode;
-            }
-            r.Close();
-            return reVal;
-
-        }
-
-
-        private static bool IsUTF8Bytes(byte[] data)
-        {
-            int charByteCounter = 1; //计算当前正分析的字符应还有的字节数
-            byte curByte; //当前分析的字节.
-            for (int i = 0; i < data.Length; i++)
-            {
-                curByte = data[i];
-                if (charByteCounter == 1)
-                {
-                    if (curByte >= 0x80)
-                    {
-                        //判断当前
-                        while (((curByte <<= 1) & 0x80) != 0)
-                        {
-                            charByteCounter++;
-                        }
-                        //标记位首位若为非0 则至少以2个1开始 如:110XXXXX…1111110X
-                        if (charByteCounter == 1 || charByteCounter > 6)
-                        {
-                            return false;
-                        }
-                    }
-                }
-                else
-                {
-                    //若是UTF-8 此时第一位必须为1
-                    if ((curByte & 0xC0) != 0x80)
-                    {
-                        return false;
-                    }
-                    charByteCounter--;
-                }
-            }
-            if (charByteCounter > 1)
-            {
-                return false;
-            }
-            return true;
-        }
-
-        private int GetContentLength(int ordinal)
-        {
-            IntPtr colLengthPrt = TDengine.FetchLengths(_taosResult);
-            int numOfFiled = TDengine.FieldCount(_taosResult);
-            int[] colLengthArr = new int[numOfFiled];
-            Marshal.Copy(colLengthPrt, colLengthArr, 0, numOfFiled);
-            return colLengthArr[ordinal];
-        }
-
+ 
         /// <summary>
         ///     Gets the value of the specified column.
         /// </summary>
         /// <param name="ordinal">The zero-based column ordinal.</param>
         /// <returns>The value of the column.</returns>
-        public override object GetValue(int ordinal)
-        {
-            object result = null;
-            if (ordinal >= 0 && ordinal < _metas.Count)
-            {
-                var meta = _metas[ordinal];
-                int offset = IntPtr.Size * ordinal;
-                IntPtr data = Marshal.ReadIntPtr(rowdata, offset);
-                if (data != IntPtr.Zero)
-                {
-                    switch ((TDengineDataType)meta.type)
-                    {
-                        case TDengineDataType.TSDB_DATA_TYPE_BOOL:
-                            bool v1 = Marshal.ReadByte(data) == 0 ? false : true;
-                            result = v1;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_TINYINT:
-                            sbyte v2s = (sbyte)Marshal.ReadByte(data);
-                            result = v2s;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_UTINYINT:
-                            byte v2 = Marshal.ReadByte(data);
-                            result = v2;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_SMALLINT:
-                            short v3 = Marshal.ReadInt16(data);
-                            result = v3;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_USMALLINT:
-                            ushort v12 = (ushort)Marshal.ReadInt16(data);
-                            result = v12;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_INT:
-                            int v4 = Marshal.ReadInt32(data);
-                            result = v4;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_UINT:
-                            uint v13 = (uint)Marshal.ReadInt32(data);
-                            result = v13;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_BIGINT:
-                            long v5 = Marshal.ReadInt64(data);
-                            result = v5;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_UBIGINT:
-                            ulong v14 = (ulong)Marshal.ReadInt64(data);
-                            result = v14;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_FLOAT:
-                            float v6 = (float)Marshal.PtrToStructure(data, typeof(float));
-                            result = v6;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_DOUBLE:
-                            double v7 = (double)Marshal.PtrToStructure(data, typeof(double));
-                            result = v7;
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_BINARY:
-                            {
-#if NET5_0_OR_GREATER
-                                string v8 = Marshal.PtrToStringUTF8(data, GetContentLength(ordinal));
-                                result = v8?.RemoveNull();
-#else
-                            byte[] buffer = new byte[GetContentLength(ordinal)];
-                            Marshal.Copy(data, buffer, 0, buffer.Length);
-                            string v8 = Encoding.UTF8.GetString(buffer);
-                            result = v8?.RemoveNull();
-#endif
-                            }
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_TIMESTAMP:
-                            {
-                                result = GetDateTimeFrom(data);
-                            }
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_JSONTAG:
-                        case TDengineDataType.TSDB_DATA_TYPE_NCHAR:
-                            {
-                                string v10 = string.Empty;
-                                int contentLength = GetContentLength(ordinal);
-                                if (contentLength > 0)// https://github.com/maikebing/Maikebing.EntityFrameworkCore.Taos/issues/99
-                                {
-                                    byte[] bf = new byte[contentLength];
-                                    Marshal.Copy(data, bf, 0, contentLength);
+        public override object GetValue(int ordinal) => _context.GetValue(ordinal);
 
-                                    if (IsUTF8Bytes(bf) || (bf[0] == 0xEF && bf[1] == 0xBB && bf[2] == 0xBF))
-                                    {
-                                        v10 = System.Text.Encoding.UTF8.GetString(bf)?.RemoveNull();
-                                    }
-                                    else
-                                    {
-                                        v10 = System.Text.Encoding.GetEncoding(936).GetString(bf)?.RemoveNull();
-                                    }
-                                }
-                                result = v10;
-                            }
-                            break;
-                        case TDengineDataType.TSDB_DATA_TYPE_NULL:
-                            result = null;
-                            break;
-                    }
-                }
-                else
-                {
-                    result =  null;
-                }
-            }
-            return result;
-        }
 
-        private DateTime GetDateTimeFrom(IntPtr data)
-        {
-            var val = Marshal.ReadInt64(data);
-            //double tsp;
-            var _dateTimePrecision = (TSDB_TIME_PRECISION)TDengine.ResultPrecision(_taosResult);
-            switch (_dateTimePrecision)
-            {
-                /*
-                * ticks为100纳秒，必须乘以10才能达到微秒级的区分度
-                * 1秒s    = 1000毫秒ms
-                * 1毫秒ms = 1000微秒us
-                * 1微秒us = 1000纳秒ns
-                * 因此， 1毫秒ms = 1000000纳秒ns = 10000ticks
-                */
-                case TSDB_TIME_PRECISION.TSDB_TIME_PRECISION_NANO:
-                    val /= 100;
-                    break;
-                case TSDB_TIME_PRECISION.TSDB_TIME_PRECISION_MICRO:
-                    val *= 10;
-                    break;
-                case TSDB_TIME_PRECISION.TSDB_TIME_PRECISION_MILLI:
-                default:
-                    val *= 10000;
-                    break;
-            }
-            var v9 = _dt1970.AddTicks(val);
-            return v9.ToLocalTime();
-        }
+
 
 
         /// <summary>
@@ -681,92 +406,7 @@ namespace IoTSharp.Data.Taos
         ///     Returns a System.Data.DataTable that describes the column metadata of the System.Data.Common.DbDataReader.
         /// </summary>
         /// <returns>A System.Data.DataTable that describes the column metadata.</returns>
-        public override DataTable GetSchemaTable()
-        {
-            var schemaTable = new DataTable("SchemaTable");
-            if (_metas != null && _metas.Count > 0)
-            {
-                var ColumnName = new DataColumn(SchemaTableColumn.ColumnName, typeof(string));
-                var ColumnOrdinal = new DataColumn(SchemaTableColumn.ColumnOrdinal, typeof(int));
-                var ColumnSize = new DataColumn(SchemaTableColumn.ColumnSize, typeof(int));
-                var NumericPrecision = new DataColumn(SchemaTableColumn.NumericPrecision, typeof(short));
-                var NumericScale = new DataColumn(SchemaTableColumn.NumericScale, typeof(short));
-
-                var DataType = new DataColumn(SchemaTableColumn.DataType, typeof(Type));
-                var DataTypeName = new DataColumn("DataTypeName", typeof(string));
-
-                var IsLong = new DataColumn(SchemaTableColumn.IsLong, typeof(bool));
-                var AllowDBNull = new DataColumn(SchemaTableColumn.AllowDBNull, typeof(bool));
-
-                var IsUnique = new DataColumn(SchemaTableColumn.IsUnique, typeof(bool));
-                var IsKey = new DataColumn(SchemaTableColumn.IsKey, typeof(bool));
-                var IsAutoIncrement = new DataColumn(SchemaTableOptionalColumn.IsAutoIncrement, typeof(bool));
-
-                var BaseCatalogName = new DataColumn(SchemaTableOptionalColumn.BaseCatalogName, typeof(string));
-                var BaseSchemaName = new DataColumn(SchemaTableColumn.BaseSchemaName, typeof(string));
-                var BaseTableName = new DataColumn(SchemaTableColumn.BaseTableName, typeof(string));
-                var BaseColumnName = new DataColumn(SchemaTableColumn.BaseColumnName, typeof(string));
-
-                var BaseServerName = new DataColumn(SchemaTableOptionalColumn.BaseServerName, typeof(string));
-                var IsAliased = new DataColumn(SchemaTableColumn.IsAliased, typeof(bool));
-                var IsExpression = new DataColumn(SchemaTableColumn.IsExpression, typeof(bool));
-
-                var columns = schemaTable.Columns;
-
-                columns.Add(ColumnName);
-                columns.Add(ColumnOrdinal);
-                columns.Add(ColumnSize);
-                columns.Add(NumericPrecision);
-                columns.Add(NumericScale);
-                columns.Add(IsUnique);
-                columns.Add(IsKey);
-                columns.Add(BaseServerName);
-                columns.Add(BaseCatalogName);
-                columns.Add(BaseColumnName);
-                columns.Add(BaseSchemaName);
-                columns.Add(BaseTableName);
-                columns.Add(DataType);
-                columns.Add(DataTypeName);
-                columns.Add(AllowDBNull);
-                columns.Add(IsAliased);
-                columns.Add(IsExpression);
-                columns.Add(IsAutoIncrement);
-                columns.Add(IsLong);
-
-                for (var i = 0; i < _metas.Count; i++)
-                {
-                    var schemaRow = schemaTable.NewRow();
-
-                    schemaRow[ColumnName] = GetName(i);
-                    schemaRow[ColumnOrdinal] = i;
-                    schemaRow[ColumnSize] = _metas[i].Size;
-                    schemaRow[NumericPrecision] = DBNull.Value;
-                    schemaRow[NumericScale] = DBNull.Value;
-                    schemaRow[BaseServerName] = _command.Connection.DataSource;
-                    var databaseName = _command.Connection.Database;
-                    schemaRow[BaseCatalogName] = databaseName;
-                    var columnName = GetName(i);
-                    schemaRow[BaseColumnName] = columnName;
-                    schemaRow[BaseSchemaName] = DBNull.Value;
-                    var tableName = string.Empty;
-                    schemaRow[BaseTableName] = tableName;
-                    schemaRow[DataType] = GetFieldType(i);
-                    schemaRow[DataTypeName] = GetDataTypeName(i);
-                    schemaRow[IsAliased] = columnName != GetName(i);
-                    schemaRow[IsExpression] = columnName == null;
-                    schemaRow[IsLong] = DBNull.Value;
-                    if (i == 0)
-                    {
-                        schemaRow[IsKey] = true;
-                        schemaRow[DataType] = GetFieldType(i);
-                        schemaRow[DataTypeName] = GetDataTypeName(i);
-                    }
-                    schemaTable.Rows.Add(schemaRow);
-                }
-
-            }
-
-            return schemaTable;
-        }
+        public override DataTable GetSchemaTable() => _context.GetSchemaTable(GetName,GetFieldType,GetDataTypeName,_command);
+      
     }
 }
